txpool namespace - porting to m3_v3
